organize so debug stuff is in debug files
